+++ test/runtests.jl
@@ -5,6 +5,7 @@ global_logger(ConsoleLogger(stderr, Error))
 
 haskey(ENV, "TRACKMATCHER_PROGRESS") || (ENV["TRACKMATCHER_PROGRESS"] = "false")
 
+include("init.jl")
 include("test_flightdata.jl")
 include("test_clouddata.jl")
 include("test_satdata.jl")