+++ test/test_satdata.jl
@@ -1,14 +1,5 @@
 ## Define expected results
-
-# Load satellite data (and reuse for intercept finding)
-cpro_src = joinpath(@__DIR__, "data", "caliop", "CPro")
-clay_src = joinpath(@__DIR__, "data", "caliop", "CLay")
-sat_cpro = SatSet(cpro_src, type=:CPro)
-sat_clay = SatSet(clay_src, type=:CLay)
-timeindex = [2035:2065]
-lidarprofile = TrackMatcher.get_lidarheights((15_000, -Inf), Float32)
-cpro = CPro([joinpath(@__DIR__, "data", "caliop", "CPro", "CPro_4.h5")], timeindex, lidarprofile)
-clay = CLay([joinpath(@__DIR__, "data", "caliop", "CLay", "CLay_4.h5")], timeindex, (15_000, -Inf))
+# ¡ Needs general test data from init.jl
 
 # Target data for sat test sets
 cpro_files = [joinpath("Level1", "CPro1.h5"), joinpath("Level1", "Level2", "CPro2.h5")]